moves some tab related code to mixin
